+++ core/output.py
@@ -67,7 +67,7 @@ class Output:
         if error is not None:
             tr = "".join(traceback.format_exception(type(error), error, error.__traceback__))
             if len(tr) > 1000:
-                tr = tr[-999:] + "…"
+                tr = "…" + tr[-999:]
             result += template_cont.format(error=error, traceback=tr)
 
         await source.send(result, delete_after=config.get("delay", "bot error"))
@@ -207,7 +207,7 @@ class Event:
         # fmt: off
         await self.getChannel().send(self.user_template.format(
             user=str(member),
-            location=location.mention,
+            location=location.mention if hasattr(location, "mention") else type(location).__name__,
             message=message.replace("@", "@\u200b"),
         ))
         # fmt: on
@@ -217,7 +217,7 @@ class Event:
         # fmt: off
         await self.getChannel().send(self.sudo_template.format(
             user=str(member),
-            location=location.mention,
+            location=location.mention if hasattr(location, "mention") else type(location).__name__,
             message=message.replace("@", "@\u200b"),
         ))
         # fmt: on